requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorlover==0.3.0
comm==0.1.3
contourpy==1.0.7
cufflinks==0.17.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.19.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.13.2
ipywidgets==8.0.6
jax==0.4.10
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.1.0
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.8
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pandasql==0.7.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.14.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
pydicom==2.3.1
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
SQLAlchemy==1.4.48
stack-data==0.6.2
streamlit==1.22.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
textblob==0.17.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.0
tzdata==2023.3
tzlocal==5.0.1
ultralytics==8.0.111
urllib3==1.26.16
validators==0.20.0
virtualenv==20.23.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
widgetsnbextension==4.0.7
wrapt==1.14.1
yarl==1.9.2
zipp==3.15.0
